refactor: split claim action panel
